Started work on staging releases.
The idea is to deploy the maven artifacts to OSS Repository Hosting
(OSSRH) once we declare the project and to deploy the source and binary
zip to our own servers.

Deploying to OSSRH will be done automatically with maven using
nexus-staging-maven-plugin but with autoReleaseAfterClose set to false.
Deploying to our servers will be done automatically with maven using the
Wagon plugin configured to a staging area. From there, the project can
be reviewed before being made available to public.