Fix Findbugs, PMD, Checkstyle. Site now works
Split the findbugs exclude fiter xml document into one for each module since
its contents in module specific.

For Checkstyle and PMD I created a new module, build-tools, that has the
required files in src/main/resources following the recomended setup for
multi-module projects from Checkstyle's website. Checkstyle and PMD now declare
a dependency on the new module in the build/plugins section of the parent pom
so that it will be include when running `mvn site`.