Updated headers for new files or files really heavily changed.
We took care to not change too much headers, as many files come
from the Apache software Foundation. However, the Hipparchus
project is composed of the original authors of most if not all of
the files and ASF does not claim copyright transfer. So these files
really can be contributed by these authors to the Hipparchus project.
We could probably have changed much more headers.