Fixed infinite loop in Gragg-Bulirsch-Stoer integrator.
The infinite loop appeared in some cases of some step reduction.
The step was silently filtered and increased back to min step without
notifying the user that the problem needed smaller steps. Then the
step was rejected, an new attempt made with a new step ... again
increased back to min step, again rejected ...