Brent solvers converge when they can't improve
When a bracketing Brent solver has reduced the interval to the point where it
can't get any smaller (there are no more numbers in between xA and xB) then
treat it as if the root finder had converged normally. Previously a max
iterations exceeded exception would be thrown because the root finder would
continue looping without making progress.

This became an issue when event finding switched to only using the user
supplied tolerance for the root finder.