Fix out of order event processing in ODE
In some circumstances ODE events were processed out of order due to the
PriorityQueue used to sort events becoming corrupted. Previously when updating
the sort order of an event it was not immediately removed from the queue which
could lead to a violation of the heap invariant. Now the updated object is
removed before taking any other action with the queue which preserves the
invariant. Unfortunately this bug is difficult to reproduce and I could not
create a reliable test case.

Fixes issue #1