Fix Field ODE interpolation with restricted step
The Field interpolators would use the state at the ends of the restricted step
instead of the ends of the global (integrator) step. This appeared to be an
error when copying the double code to use Field. Also copied tests from the
double versions to the Field version to check this issue. Had to update the
tolerance values in LutherFieldIntegratorTest to match those in
LutherIntegratorTest.

Fixes #33