Removed Stream API by default.
Streams are great, but they impose some restrictions that can be
difficult to overcome. Kalman filter is stateful, and user processes
may generate exceptions. This does not suit well in the Streams
framework.

If users want to use streams (and in fact we do this in the Hipparchus
unit tests), they have to take care by themselves of the caveats.