Don't separate compilation from test generation.
The test-jar created in core module is reused in the tests of
other modules and therefore is a dependency with scope test.
However, due to maven way of handling multi-module packages, this
dependency *must* exist even before we attempt to run tests, so
we cannot compile everything without tests and then test everything.
We must compile and test core, then compile and test module 1, then
compile and test module 2...