Prepared tests for future change to allow Complex differentiation.
The abs() conflict between Complex and realfieldElement() currently
prevents Complex to be used in FieldDerivativeStructure. Once this
conflict is solved (which introduces an incompatibility and cannot be
done in the 1.X series), then we will be able to add
FieldDerivativeStructureComplexTest (there is already such an
implementation, but not pushed to the master branch as of 2020-06-02.